🗑️ Deprecate `biopipen.namespaces`, use `biopipen.ns` instead
