+++ biopipen/namespaces/__init__.py
@@ -1,0 +1,7 @@
+import warnings
+
+warnings.simplefilter('always', DeprecationWarning)
+warnings.warn(
+    "`biopipen.namespaces` is deprecated. Please use `biopipen.ns` instead.",
+    DeprecationWarning,
+)
